move common.py up
