+++ tools/visual-pipeline-and-platform-evaluation-tool/pipelines/common.py
